pyproject.toml
[tool.poetry]
name = "crew"
version = "0.1.0"
description = ""
authors = ["Jack Dewinter <jackdewinter1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<=3.13"
crewai = {extras = ["tools"], version = "^0.30.11"}

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
